Renames tdmprograms
