+++ strawberryfields/tdm/tdmprogram.py
@@ -1,4 +1,4 @@
-# Copyright 2029 Xanadu Quantum Technologies Inc.
+# Copyright 2019 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -47,7 +47,7 @@ def get_modes(cmd, q):
     return modes
 
 
-def validate_measurements(circuit, shift):
+def validate_measurements(circuit, shift, N):
     """Validate the TDM program measurements are correct"""
     spatial_modes = 0
     measurement_reg = []
@@ -67,6 +67,13 @@ def validate_measurements(circuit, shift):
     # if shift == "after" and list(set(measurement_reg)) != [0]:
     #     raise ValueError("Measurements allowed only on q[0].")
 
+    #####################################################################################
+    ##### THIS CHANGE ACCOUNTS FOR NEW PROPOSAL: ONE BELT PER SPATIAL MODE ##############
+    #####################################################################################
+    # N is now a list with number of qumodes for each spatial mode
+    if spatial_modes is not len(N):
+        raise ValueError("Numer of measurements operators must match number of spatial_modes.")
+
     return spatial_modes
 
 
@@ -83,8 +90,7 @@ def input_check(args, copies):
 
 
 def reshape_samples(samples, c=None):
-    """Reshapes the samples so that they have the expected correct shape.
-    """
+    """Reshapes the samples so that they have the expected correct shape"""
     num_modes = len(samples)
     len_samples = [len(samples[i]) for i in range(num_modes)]
     max_len = max(len_samples)
@@ -99,18 +105,24 @@ def reshape_samples(samples, c=None):
         return reshaped_samples
     return reshaped_samples.reshape(c, -1)
 
-def reshape_samples_with_copies(samples, copies, spatial_modes):
-    reshaped_samples = reshape_samples(samples)
-    return reshaped_samples.reshape([copies, spatial_modes, -1])
 
 class TDMProgram(sf.Program):
     """Represents a photonic quantum circuit in the time domain encoding"""
 
     def __init__(self, N, name=None):
-        self.concurr_modes = N
+        #####################################################################################
+        ##### THIS CHANGE ACCOUNTS FOR NEW PROPOSAL: ONE BELT PER SPATIAL MODE ##############
+        #####################################################################################
+        # N is now a list with number of qumodes for each spatial mode
+        if isinstance(N, int):
+            self.N = [N]
+        else:
+            self.N = N
+        self.concurr_modes = sum(self.N)
+
         self.total_timebins = 0
         self.spatial_modes = 0
-        super().__init__(num_subsystems=N, name=name)
+        super().__init__(num_subsystems=self.concurr_modes, name=name)
 
     def context(self, *args, copies=1, shift="end"):
         input_check(args, copies)
@@ -126,7 +138,7 @@ class TDMProgram(sf.Program):
 
     def __exit__(self, ex_type, ex_value, ex_tb):
         super().__exit__(ex_type, ex_value, ex_tb)
-        self.spatial_modes = validate_measurements(self.circuit, self.shift)
+        self.spatial_modes = validate_measurements(self.circuit, self.shift, self.N)
         self.construct_circuit()
 
     def construct_circuit(self):
@@ -139,6 +151,41 @@ class TDMProgram(sf.Program):
 
         q = self.register
 
+
+        #####################################################################################
+        ##### THIS CHANGE ACCOUNTS FOR NEW PROPOSAL: ONE BELT PER SPATIAL MODE ##############
+        #####################################################################################
+        # N is now a list with number of qumodes for each spatial mode
+
+        sm = []
+        for i in range(len(self.N)):
+            start = sum(self.N[:i])
+            stop = sum(self.N[:i])+self.N[i]
+            sm.append(slice(start, stop))
+        ''' ^ list of slice intervals; each slice interval corresponding to one spatial mode
+
+        For instance, say
+        N = [5,4,9,1],
+        then this corresponds to
+        5 concurrent modes in spatial mode A
+        4 concurrent modes in spatial mode B
+        9 concurrent modes in spatial mode C
+        1 concurrent modes in spatial mode D.
+    
+        The entries of sm will then be:
+        slice(0, 6, None)   for spatial mode A
+        slice(6, 10, None)  for spatial mode B
+        slice(10, 19, None) for spatial mode C
+        slice(19, 20, None) for spatial mode D.
+
+        This helps to define:
+        q[sm[0]] as concurrent modes of spatial mode A
+        q[sm[1]] as concurrent modes of spatial mode B
+        q[sm[2]] as concurrent modes of spatial mode C
+        q[sm[3]] as concurrent modes of spatial mode D.
+    
+        '''
+
         for i in range(self.total_timebins):
             for cmd in cmds:
                 self.apply_op(cmd, q, i)
@@ -147,7 +194,15 @@ class TDMProgram(sf.Program):
                     q = shift_by(q, 1)  # shift after each measurement
 
             if self.shift == "end":
-                q = shift_by(q, self.spatial_modes)  # shift at end of each time bin
+                #####################################################################################
+                ##### THIS CHANGE ACCOUNTS FOR NEW PROPOSAL: ONE BELT PER SPATIAL MODE ##############
+                #####################################################################################
+                # shift each spatial mode SEPARATELY by one step
+                q_aux = list(q)
+                for i in range(len(self.N)):
+                    q_aux[sm[i]] = shift_by(q_aux[sm[i]], 1)
+                q = tuple(q_aux)
+               
             elif isinstance(self.shift, int):
                 q = shift_by(q, self.shift)  # shift at end of each time bin
